Fix Windows build: cast TerminationStatus.Code to Int32
On Windows, TerminationStatus.Code is UInt32, not Int32.

+++ Tests/TutorialTests/TestHelpers.swift
@@ -35,7 +35,7 @@ struct SubprocessResult: Sendable {
     var exitCode: Int32 {
         switch terminationStatus {
         case .exited(let code):
-            return code
+            return Int32(code)
         case .unhandledException:
             return -1
         @unknown default:
